refactor: Make GUI and compiler use same function to generate node names
